Decompose window manager
