--- src/window_manager.rs
@@ -1,2997 +1,0 @@
-use crate::api::WindowsApi;
-use crate::common::*;
-use crate::configuration_provider::{
-  ADDITIONAL_WORKSPACE_COUNT, ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE, ALLOW_SELECTING_SAME_CENTER_WINDOWS,
-  ConfigurationProvider, Layout, SCROLLING_ANIMATION_DURATION_IN_MS, WINDOW_MARGIN,
-};
-use crate::horizontal_layout::HorizontalLayout;
-use crate::utils::{
-  CONFIGURATION_PROVIDER_LOCK, MINIMUM_WINDOW_DIMENSION, MINIMUM_WINDOW_DIMENSION_DIVISOR, MINIMUM_WINDOW_MARGIN,
-};
-use crate::workspace_manager::WorkspaceManager;
-use std::collections::HashMap;
-use std::sync::{Arc, Mutex};
-use std::time::Duration;
-use windows::Win32::UI::Shell::IVirtualDesktopManager;
-use windows::Win32::UI::WindowsAndMessaging::SW_MAXIMIZE;
-
-const REGULAR_TOLERANCE_IN_PX: i32 = 2;
-const DWM_TOLERANCE_IN_PX: i32 = 8;
-const HORIZONTAL_ANIMATION_FRAMES: u32 = 12;
-
-pub struct WindowManager<T: WindowsApi> {
-  configuration_provider: Arc<Mutex<ConfigurationProvider>>,
-  known_windows: HashMap<String, WindowPlacement>,
-  allow_moving_cursor_after_close_or_minimise: bool,
-  horizontal_layout: HorizontalLayout,
-  horizontal_layout_positions: HashMap<PersistentWorkspaceId, Vec<(WindowHandle, Rect)>>,
-  horizontal_layout_initialised: bool,
-  previous_foreground_window: Option<WindowHandle>,
-  workspace_manager: WorkspaceManager<T>,
-  virtual_desktop_manager: Option<IVirtualDesktopManager>,
-  windows_api: T,
-}
-
-impl<T: WindowsApi + Clone> WindowManager<T> {
-  pub fn new(configuration_provider: Arc<Mutex<ConfigurationProvider>>, api: T) -> Self {
-    let guard = match configuration_provider.try_lock() {
-      Ok(guard) => guard,
-      Err(err) => {
-        panic!(
-          "{} when trying to create window manager: {}",
-          CONFIGURATION_PROVIDER_LOCK, err
-        );
-      }
-    };
-    let additional_workspace_count = guard.get_i32(ADDITIONAL_WORKSPACE_COUNT);
-    let window_margin = guard.get_i32(WINDOW_MARGIN);
-    let allow_moving_cursor_after_close_or_minimise = guard.get_bool(ALLOW_MOVING_CURSOR_AFTER_OPEN_CLOSE_OR_MINIMISE);
-    drop(guard);
-    let workspace_manager = WorkspaceManager::new(additional_workspace_count, window_margin, api.clone());
-
-    Self {
-      known_windows: HashMap::new(),
-      allow_moving_cursor_after_close_or_minimise,
-      horizontal_layout: HorizontalLayout::default(),
-      horizontal_layout_positions: HashMap::new(),
-      horizontal_layout_initialised: false,
-      previous_foreground_window: None,
-      virtual_desktop_manager: Some(
-        api
-          .get_virtual_desktop_manager()
-          .expect("Failed to get the virtual desktop manager"),
-      ),
-      workspace_manager,
-      configuration_provider,
-      windows_api: api,
-    }
-  }
-
-  /// Returns the unique IDs for all desktop containers across all monitors in their natural order.
-  pub fn get_ordered_permanent_workspace_ids(&mut self) -> Vec<PersistentWorkspaceId> {
-    self.workspace_manager.get_ordered_permanent_workspace_ids()
-  }
-
-  pub fn close_window(&mut self) {
-    let Some(window) = self.windows_api.get_foreground_window() else {
-      return;
-    };
-
-    if self.layout_for_window(window) == Some(Layout::Scrolling) {
-      self.windows_api.do_close_window(window);
-      self.remove_horizontal_member_and_refocus(window);
-      return;
-    }
-
-    self.windows_api.do_close_window(window);
-    if self.allow_moving_cursor_after_close_or_minimise {
-      self.find_and_select_closest_window(window);
-    }
-  }
-
-  pub fn switch_workspace(&mut self, id: PersistentWorkspaceId) {
-    if self.layout_for_workspace(id) != Some(Layout::Scrolling) {
-      self.workspace_manager.switch_workspace(id);
-      return;
-    }
-
-    let source = self
-      .workspace_manager
-      .active_workspace_ids()
-      .into_iter()
-      .find(|workspace| workspace.monitor_id == id.monitor_id);
-    let additional_windows = source.map_or_else(Vec::new, |workspace| self.horizontal_layout.members(workspace).to_vec());
-    self
-      .workspace_manager
-      .switch_workspace_with_additional_windows(id, &additional_windows);
-    self.reflow_horizontal_workspace(id);
-    self.focus_horizontal_workspace(id);
-  }
-
-  pub fn move_window_to_workspace(&mut self, id: PersistentWorkspaceId) {
-    let foreground = self.windows_api.get_foreground_window();
-    let source = foreground.and_then(|handle| self.workspace_for_window(handle));
-    if source == Some(id) || self.workspace_manager.monitor_for_workspace(id).is_none() {
-      return;
-    }
-    let source_layout = source.and_then(|workspace| self.layout_for_workspace(workspace));
-    let target_layout = self.layout_for_workspace(id);
-    self.workspace_manager.move_window_to_workspace(id);
-
-    if let (Some(handle), Some(source)) = (foreground, source) {
-      if source_layout == Some(Layout::Scrolling) {
-        self.horizontal_layout.remove(source, handle);
-      }
-      if target_layout == Some(Layout::Scrolling) {
-        self.horizontal_layout.insert_before(id, handle, None);
-      }
-      if source_layout == Some(Layout::Scrolling) && self.workspace_manager.is_workspace_active(source) {
-        self.reflow_horizontal_workspace(source);
-        self.focus_horizontal_workspace(source);
-      }
-      if target_layout == Some(Layout::Scrolling) && self.workspace_manager.is_workspace_active(id) {
-        self.reflow_horizontal_workspace(id);
-        self.focus_horizontal_workspace(id);
-      }
-    }
-  }
-
-  pub fn move_window(&mut self, direction: Direction) {
-    if self.foreground_layout() == Some(Layout::Scrolling) {
-      if matches!(direction, Direction::Left | Direction::Right) {
-        self.reorder_horizontal_window(direction);
-      }
-      return;
-    }
-    let (handle, placement, monitor_info) = match self.get_window_and_monitor_info() {
-      Some(value) => value,
-      None => return,
-    };
-    let sizing = match direction {
-      Direction::Left => Sizing::left_half_of_screen(monitor_info.work_area, self.margin()),
-      Direction::Right => Sizing::right_half_of_screen(monitor_info.work_area, self.margin()),
-      Direction::Up => Sizing::top_half_of_screen(monitor_info.work_area, self.margin()),
-      Direction::Down => Sizing::bottom_half_of_screen(monitor_info.work_area, self.margin()),
-    };
-
-    match self.is_of_expected_size(handle, &placement, &sizing) {
-      true => {
-        let all_monitors = self.windows_api.get_all_monitors();
-        let this_monitor = self.windows_api.get_monitor_handle_for_window_handle(handle);
-        let target_monitor = all_monitors.get(direction, this_monitor);
-        if let Some(target_monitor) = target_monitor {
-          debug!("Moving window to [{}]", target_monitor);
-          self.windows_api.set_window_position(handle, target_monitor.work_area);
-          self.near_maximise_window(handle, MonitorInfo::from(target_monitor), self.margin());
-          self.windows_api.set_cursor_position(&target_monitor.center);
-        } else {
-          debug!("No monitor found in [{:?}] direction, did not move window", direction);
-        }
-      }
-      false => {
-        let cursor_target_point = Point::from_center_of_sizing(&sizing);
-        self.execute_window_resizing(handle, sizing);
-        self.windows_api.set_cursor_position(&cursor_target_point);
-      }
-    }
-  }
-
-  pub fn resize_window(&mut self, direction: Direction) {
-    if self.foreground_layout() == Some(Layout::Scrolling) {
-      return;
-    }
-
-    let (handle, placement, monitor_info) = match self.get_window_and_monitor_info() {
-      Some(value) => value,
-      None => return,
-    };
-    let work_area = monitor_info.work_area;
-    let current_sizing = Sizing::from(placement.normal_position);
-    let new_sizing = if self.is_near_maximised(&placement, &handle, &monitor_info) {
-      Sizing::three_quarter_near_maximised(work_area, direction, self.margin())
-    } else if self.is_three_quarter_near_maximised(&handle, &monitor_info, direction) {
-      Sizing::near_maximised(work_area, self.margin()).halved(direction, self.margin())
-    } else if self.is_three_quarter_near_maximised(&handle, &monitor_info, direction.opposite()) {
-      Sizing::centre_near_maximised(work_area, direction, self.margin())
-    } else {
-      current_sizing.halved(direction, self.margin())
-    };
-    debug!(
-      "Expected size of {}: ({},{})x({},{})",
-      handle, new_sizing.x, new_sizing.y, new_sizing.width, new_sizing.height
-    );
-    let min_width = MINIMUM_WINDOW_DIMENSION.max((work_area.right - work_area.left) / MINIMUM_WINDOW_DIMENSION_DIVISOR);
-    let min_height = MINIMUM_WINDOW_DIMENSION.max((work_area.bottom - work_area.top) / MINIMUM_WINDOW_DIMENSION_DIVISOR);
-    if new_sizing.width < min_width || new_sizing.height < min_height {
-      let is_below_constant = min_width <= MINIMUM_WINDOW_DIMENSION || min_height <= MINIMUM_WINDOW_DIMENSION;
-      debug!(
-        "Not resizing {} because resulting size ({}x{}) is below minimum ({}x{}) (hit {} threshold)",
-        handle,
-        new_sizing.width,
-        new_sizing.height,
-        min_width,
-        min_height,
-        if is_below_constant { "constant" } else { "dynamic" }
-      );
-      return;
-    }
-    let cursor_target_point = Point::from_center_of_sizing(&new_sizing);
-    self.execute_window_resizing(handle, new_sizing);
-    self.windows_api.set_cursor_position(&cursor_target_point);
-  }
-
-  pub fn move_cursor(&mut self, direction: Direction) {
-    if matches!(direction, Direction::Left | Direction::Right) && self.move_horizontal_focus(direction) {
-      return;
-    }
-
-    let windows = self.windows_api.get_all_visible_windows();
-    let cursor_position = self.windows_api.get_cursor_position();
-    let (ref_point, ref_window) = match self.find_window_at_cursor(&cursor_position, &windows) {
-      Some(window_info) => (Point::from_center_of_rect(&window_info.rect), Some(window_info)),
-      None => (cursor_position, None),
-    };
-
-    if let Some(target_window) = if let Some(vdm) = &self.virtual_desktop_manager {
-      self.find_closest_window_in_direction(&ref_point, direction, &windows, vdm, ref_window)
-    } else {
-      None
-    } {
-      let target_point = Point::from_center_of_rect(&target_window.rect);
-      self.move_focus_to_window(direction, target_window, &target_point);
-    } else {
-      trace!("No window found in [{:?}] direction, attempting to find monitor", direction);
-      let all_monitors = self.windows_api.get_all_monitors();
-      let this_monitor = self.windows_api.get_monitor_handle_for_point(&cursor_position);
-      match all_monitors.get(direction, this_monitor) {
-        Some(target_monitor) => self.move_focus_to_monitor(direction, target_monitor),
-        None => {
-          info!(
-            "No window or monitor found in [{:?}] direction, did not move cursor",
-            direction
-          );
-        }
-      }
-    };
-  }
-
-  pub fn near_maximise_or_restore(&mut self) {
-    let (handle, placement, monitor_info) = match self.get_window_and_monitor_info() {
-      Some(value) => value,
-      None => return,
-    };
-
-    match self.is_near_maximised(&placement, &handle, &monitor_info) {
-      true => self.restore_previous_placement(&self.known_windows, handle),
-      false => {
-        add_or_update_previous_placement(&mut self.known_windows, handle, placement);
-        self.near_maximise_window(handle, monitor_info, self.margin());
-      }
-    }
-  }
-
-  pub fn minimise_window(&mut self) {
-    let Some(window) = self.windows_api.get_foreground_window() else {
-      return;
-    };
-
-    if self.layout_for_window(window) == Some(Layout::Scrolling) {
-      self.windows_api.do_minimise_window(window);
-      self.remove_horizontal_member_and_refocus(window);
-      return;
-    }
-
-    self.windows_api.do_minimise_window(window);
-    if self.allow_moving_cursor_after_close_or_minimise {
-      self.find_and_select_closest_window(window);
-    }
-  }
-
-  pub fn restore_all_managed_windows(&mut self) {
-    self.workspace_manager.restore_all_managed_windows();
-    self.restore_off_screen_horizontal_windows();
-  }
-
-  /// Reconciles visible windows with active horizontal strips.
-  pub fn reconcile_horizontal_layout(&mut self) {
-    let active_workspaces = self
-      .workspace_manager
-      .active_workspace_ids()
-      .into_iter()
-      .filter(|workspace| self.layout_for_workspace(*workspace) == Some(Layout::Scrolling))
-      .collect::<Vec<_>>();
-    if active_workspaces.is_empty() {
-      return;
-    }
-
-    let foreground = self.windows_api.get_foreground_window();
-    if foreground.is_some_and(|handle| self.windows_api.is_not_a_managed_window(&handle)) {
-      return;
-    }
-    let previous_workspace = self
-      .previous_foreground_window
-      .and_then(|handle| self.horizontal_layout.workspace_containing(handle));
-    let mut windows = self
-      .windows_api
-      .get_all_visible_windows()
-      .into_iter()
-      .filter(|window| {
-        self
-          .virtual_desktop_manager
-          .as_ref()
-          .is_none_or(|vdm| self.windows_api.is_window_on_current_desktop(vdm, window))
-      })
-      .collect::<Vec<_>>();
-    windows.sort_by_key(|window| (window.rect.left, window.rect.top, window.handle.hwnd));
-
-    let mut by_workspace: HashMap<PersistentWorkspaceId, Vec<Window>> = HashMap::new();
-    for window in windows {
-      let workspace = self
-        .horizontal_layout
-        .workspace_containing(window.handle)
-        .filter(|workspace| self.workspace_manager.is_workspace_active(*workspace))
-        .or_else(|| self.workspace_manager.active_workspace_for_window(window.handle));
-      if let Some(workspace) = workspace
-        && self.layout_for_workspace(workspace) == Some(Layout::Scrolling)
-      {
-        by_workspace.entry(workspace).or_default().push(window);
-      }
-    }
-
-    for (workspace, windows) in &by_workspace {
-      for window in windows {
-        if let Some(previous_workspace) = self.horizontal_layout.workspace_containing(window.handle)
-          && previous_workspace != *workspace
-        {
-          self.horizontal_layout.remove(previous_workspace, window.handle);
-        }
-      }
-    }
-
-    let mut newly_focused = None;
-    if !self.horizontal_layout_initialised {
-      for workspace in &active_workspaces {
-        let members = by_workspace
-          .get(workspace)
-          .map(|windows| windows.iter().map(|window| window.handle).collect())
-          .unwrap_or_default();
-        self.horizontal_layout.adopt(*workspace, members, foreground);
-      }
-      self.horizontal_layout_initialised = true;
-    } else {
-      for workspace in &active_workspaces {
-        let visible = by_workspace
-          .get(workspace)
-          .map(|windows| windows.iter().map(|window| window.handle).collect::<Vec<_>>())
-          .unwrap_or_default();
-        self.horizontal_layout.retain(*workspace, &visible);
-
-        let new_members = visible
-          .iter()
-          .copied()
-          .filter(|handle| self.horizontal_layout.workspace_containing(*handle).is_none())
-          .collect::<Vec<_>>();
-        let desired_focus = foreground
-          .filter(|handle| new_members.contains(handle))
-          .or_else(|| new_members.last().copied());
-        for member in new_members {
-          self
-            .horizontal_layout
-            .insert_before(*workspace, member, self.previous_foreground_window);
-        }
-        if let Some(member) = desired_focus {
-          self.horizontal_layout.set_focused(*workspace, member);
-          newly_focused = Some((*workspace, member));
-        }
-      }
-    }
-
-    for workspace in active_workspaces {
-      self.reflow_horizontal_workspace(workspace);
-    }
-    if let Some((workspace, _)) = newly_focused {
-      self.reflow_horizontal_workspace(workspace);
-      self.focus_horizontal_workspace(workspace);
-    } else if let Some(handle) = foreground
-      && let Some(workspace) = self.horizontal_layout.workspace_containing(handle)
-    {
-      self.horizontal_layout.set_focused(workspace, handle);
-      self.reflow_horizontal_workspace(workspace);
-    } else if let Some(workspace) = previous_workspace
-      && self.workspace_manager.is_workspace_active(workspace)
-      && self
-        .previous_foreground_window
-        .is_some_and(|handle| self.horizontal_layout.workspace_containing(handle).is_none())
-    {
-      self.reflow_horizontal_workspace(workspace);
-      self.focus_horizontal_workspace(workspace);
-    }
-    self.previous_foreground_window = self.windows_api.get_foreground_window();
-  }
-
-  fn layout_for_monitor(&self, monitor: &Monitor) -> Layout {
-    self
-      .configuration_provider
-      .lock()
-      .expect(CONFIGURATION_PROVIDER_LOCK)
-      .layout_for_monitor(&monitor.id_to_string(), monitor.is_primary)
-  }
-
-  fn layout_for_workspace(&self, workspace: PersistentWorkspaceId) -> Option<Layout> {
-    self
-      .workspace_manager
-      .monitor_for_workspace(workspace)
-      .map(|monitor| self.layout_for_monitor(&monitor))
-  }
-
-  fn workspace_for_window(&self, window: WindowHandle) -> Option<PersistentWorkspaceId> {
-    self
-      .horizontal_layout
-      .workspace_containing(window)
-      .or_else(|| self.workspace_manager.active_workspace_for_window(window))
-  }
-
-  fn layout_for_window(&self, window: WindowHandle) -> Option<Layout> {
-    self
-      .workspace_for_window(window)
-      .and_then(|workspace| self.layout_for_workspace(workspace))
-  }
-
-  fn foreground_layout(&self) -> Option<Layout> {
-    self
-      .windows_api
-      .get_foreground_window()
-      .and_then(|window| self.layout_for_window(window))
-  }
-
-  fn restore_off_screen_horizontal_windows(&self) {
-    let monitors = self.windows_api.get_all_monitors();
-    let screen_areas = monitors
-      .get_all()
-      .into_iter()
-      .map(|monitor| monitor.monitor_area)
-      .collect::<Vec<_>>();
-    let fallback_monitor = monitors
-      .get_all()
-      .into_iter()
-      .find(|monitor| monitor.is_primary)
-      .or_else(|| monitors.get_all().into_iter().next());
-    let margin = self.margin();
-
-    for workspace in self.horizontal_layout.workspace_ids() {
-      let Some(monitor) = monitors.get_by_id(&workspace.monitor_id).or(fallback_monitor) else {
-        continue;
-      };
-      let target = Rect::from(Sizing::near_maximised(monitor.work_area, margin));
-      for handle in self.horizontal_layout.members(workspace) {
-        let is_off_screen = self
-          .windows_api
-          .get_window_rect(*handle)
-          .is_some_and(|rect| !screen_areas.iter().any(|area| rect.intersects(area)));
-        if is_off_screen {
-          self.windows_api.set_window_position(*handle, target);
-        }
-      }
-    }
-  }
-
-  fn reflow_horizontal_workspace(&mut self, workspace: PersistentWorkspaceId) {
-    let Some(monitor) = self.workspace_manager.monitor_for_workspace(workspace) else {
-      return;
-    };
-    let positions = self
-      .horizontal_layout
-      .placements(workspace, monitor.work_area, self.margin())
-      .into_iter()
-      .map(|(handle, sizing)| (handle, Rect::from(sizing)))
-      .collect::<Vec<_>>();
-    if self.horizontal_layout_positions.get(&workspace) == Some(&positions) {
-      return;
-    }
-    if let Some(focused) = self.horizontal_layout.focused(workspace) {
-      self.windows_api.set_window_positions(&positions, focused);
-    }
-    self.horizontal_layout_positions.insert(workspace, positions);
-  }
-
-  fn animate_horizontal_workspace(&mut self, workspace: PersistentWorkspaceId, outgoing: WindowHandle) {
-    let Some(monitor) = self.workspace_manager.monitor_for_workspace(workspace) else {
-      return;
-    };
-    let desired = self
-      .horizontal_layout
-      .placements(workspace, monitor.work_area, self.margin())
-      .into_iter()
-      .map(|(handle, sizing)| (handle, Rect::from(sizing)))
-      .collect::<Vec<_>>();
-    let starts = desired
-      .iter()
-      .map(|(handle, target)| (*handle, self.windows_api.get_window_rect(*handle).unwrap_or(*target)))
-      .collect::<HashMap<_, _>>();
-    let animation_duration_in_ms = self
-      .configuration_provider
-      .lock()
-      .expect(CONFIGURATION_PROVIDER_LOCK)
-      .get_i32(SCROLLING_ANIMATION_DURATION_IN_MS);
-    let animation_duration = Duration::from_millis(u64::try_from(animation_duration_in_ms).unwrap_or_default());
-    let frame_duration = animation_duration / HORIZONTAL_ANIMATION_FRAMES;
-    for frame in 1..=HORIZONTAL_ANIMATION_FRAMES {
-      let progress = f64::from(frame) / f64::from(HORIZONTAL_ANIMATION_FRAMES);
-      let eased_progress = 1.0 - (1.0 - progress).powi(3);
-      let positions = desired
-        .iter()
-        .map(|(handle, target)| {
-          let start = starts.get(handle).copied().unwrap_or(*target);
-          (*handle, interpolate_rect(start, *target, eased_progress))
-        })
-        .collect::<Vec<_>>();
-      self.windows_api.set_window_positions(&positions, outgoing);
-      if frame < HORIZONTAL_ANIMATION_FRAMES {
-        std::thread::sleep(frame_duration);
-      }
-    }
-    if let Some(focused) = self.horizontal_layout.focused(workspace) {
-      self.windows_api.set_window_positions(&desired, focused);
-    }
-    self.horizontal_layout_positions.insert(workspace, desired);
-  }
-
-  fn focus_horizontal_workspace(&self, workspace: PersistentWorkspaceId) {
-    let Some(handle) = self.horizontal_layout.focused(workspace) else {
-      return;
-    };
-    let Some(monitor) = self.workspace_manager.monitor_for_workspace(workspace) else {
-      return;
-    };
-    let sizing = Sizing::near_maximised(monitor.work_area, self.margin());
-    self.windows_api.set_foreground_window(handle);
-    self.windows_api.set_cursor_position(&Point::from_center_of_sizing(&sizing));
-  }
-
-  fn move_horizontal_focus(&mut self, direction: Direction) -> bool {
-    let Some(current) = self.windows_api.get_foreground_window() else {
-      return false;
-    };
-    let Some(workspace) = self.horizontal_layout.workspace_containing(current) else {
-      return false;
-    };
-    if self.horizontal_layout.focus_adjacent(workspace, current, direction).is_some() {
-      self.animate_horizontal_workspace(workspace, current);
-      self.focus_horizontal_workspace(workspace);
-    }
-    true
-  }
-
-  fn reorder_horizontal_window(&mut self, direction: Direction) {
-    let Some(current) = self.windows_api.get_foreground_window() else {
-      return;
-    };
-    let Some(workspace) = self.horizontal_layout.workspace_containing(current) else {
-      return;
-    };
-    if self.horizontal_layout.reorder(workspace, current, direction) {
-      self.reflow_horizontal_workspace(workspace);
-      self.focus_horizontal_workspace(workspace);
-    }
-  }
-
-  fn remove_horizontal_member_and_refocus(&mut self, member: WindowHandle) {
-    let Some(workspace) = self.horizontal_layout.workspace_containing(member) else {
-      return;
-    };
-    self.horizontal_layout.remove(workspace, member);
-    self.reflow_horizontal_workspace(workspace);
-    self.focus_horizontal_workspace(workspace);
-  }
-
-  fn margin(&self) -> i32 {
-    let margin = self.configuration_provider.lock().unwrap().get_i32(WINDOW_MARGIN);
-    if margin >= MINIMUM_WINDOW_MARGIN { margin } else { 0 }
-  }
-
-  fn restore_previous_placement(&self, known_windows: &HashMap<String, WindowPlacement>, handle: WindowHandle) {
-    let window_id = format!("{:?}", handle.hwnd);
-    if let Some(previous_placement) = known_windows.get(&window_id) {
-      info!("Restoring previous placement for {}", window_id);
-      self
-        .windows_api
-        .do_restore_window_placement(handle, previous_placement.clone());
-    } else {
-      warn!("No previous placement found for {}", window_id);
-    }
-  }
-
-  fn is_near_maximised(&self, placement: &WindowPlacement, handle: &WindowHandle, monitor_info: &MonitorInfo) -> bool {
-    if placement.show_cmd == SW_MAXIMIZE.0 as u32 && self.margin() < MINIMUM_WINDOW_MARGIN {
-      debug!("{} is reported as maximised and margins are disabled", handle);
-      return true;
-    }
-
-    let work_area = monitor_info.work_area;
-    let expected_x = work_area.left + self.margin();
-    let expected_y = work_area.top + self.margin();
-    let expected_width = work_area.right - work_area.left - self.margin() * 2;
-    let expected_height = work_area.bottom - work_area.top - self.margin() * 2;
-    let sizing = Sizing::new(expected_x, expected_y, expected_width, expected_height);
-
-    if let Some(rect) = self.windows_api.get_window_rect(*handle) {
-      let result = (rect.left - expected_x).abs() <= REGULAR_TOLERANCE_IN_PX
-        && (rect.top - expected_y).abs() <= REGULAR_TOLERANCE_IN_PX
-        && (rect.right - rect.left - expected_width).abs() <= REGULAR_TOLERANCE_IN_PX
-        && (rect.bottom - rect.top - expected_height).abs() <= REGULAR_TOLERANCE_IN_PX;
-
-      log_actual_vs_expected(handle, &sizing, rect);
-      debug!(
-        "{} {} near-maximised (tolerance: {})",
-        handle,
-        if result { "is currently" } else { "is currently NOT" },
-        REGULAR_TOLERANCE_IN_PX
-      );
-
-      result
-    } else {
-      warn!("{} has no window rect, assuming currently NOT near-maximised", handle);
-      false
-    }
-  }
-
-  fn is_three_quarter_near_maximised(
-    &self,
-    handle: &WindowHandle,
-    monitor_info: &MonitorInfo,
-    direction: Direction,
-  ) -> bool {
-    let expected = Sizing::three_quarter_near_maximised(monitor_info.work_area, direction, self.margin());
-
-    if let Some(rect) = self.windows_api.get_window_rect(*handle) {
-      let result = (rect.left - expected.x).abs() <= REGULAR_TOLERANCE_IN_PX
-        && (rect.top - expected.y).abs() <= REGULAR_TOLERANCE_IN_PX
-        && (rect.right - rect.left - expected.width).abs() <= REGULAR_TOLERANCE_IN_PX
-        && (rect.bottom - rect.top - expected.height).abs() <= REGULAR_TOLERANCE_IN_PX;
-      debug!(
-        "{} {} three-quarter near-maximised in [{:?}] direction (tolerance: {})",
-        handle,
-        if result { "is currently" } else { "is currently NOT" },
-        direction,
-        REGULAR_TOLERANCE_IN_PX
-      );
-      result
-    } else {
-      warn!(
-        "{} has no window rect, assuming currently NOT three-quarter near-maximised",
-        handle
-      );
-      false
-    }
-  }
-
-  fn near_maximise_window(&self, handle: WindowHandle, monitor_info: MonitorInfo, margin: i32) {
-    info!("Near-maximising {}", handle);
-
-    // First maximise to get the animation effect
-    self.windows_api.do_maximise_window(handle);
-
-    // Then resize the window to the expected size
-    if margin >= MINIMUM_WINDOW_MARGIN {
-      let work_area = monitor_info.work_area;
-      let sizing = Sizing::near_maximised(work_area, margin);
-      self.execute_window_resizing(handle, sizing);
-    }
-  }
-
-  fn find_closest_window_in_direction<'window>(
-    &self,
-    reference_point: &Point,
-    direction: Direction,
-    windows: &'window [Window],
-    virtual_desktop_manager: &IVirtualDesktopManager,
-    reference_window: Option<&Window>,
-  ) -> Option<&'window Window> {
-    // Keep only current-desktop windows
-    let windows_on_current_desktop = windows
-      .iter()
-      .filter(|window| self.windows_api.is_window_on_current_desktop(virtual_desktop_manager, window))
-      .filter(|window| {
-        self
-          .horizontal_layout
-          .workspace_containing(window.handle)
-          .is_none_or(|workspace| self.horizontal_layout.focused(workspace) == Some(window.handle))
-      })
-      .collect::<Vec<_>>();
-    let allow_selecting_same_center_windows = self
-      .configuration_provider
-      .lock()
-      .expect(CONFIGURATION_PROVIDER_LOCK)
-      .get_bool(ALLOW_SELECTING_SAME_CENTER_WINDOWS);
-
-    select_window_in_direction(
-      reference_point,
-      direction,
-      &windows_on_current_desktop,
-      reference_window,
-      allow_selecting_same_center_windows,
-    )
-  }
-
-  fn execute_window_resizing(&self, handle: WindowHandle, sizing: Sizing) {
-    let placement = WindowPlacement::new_from_sizing(sizing.clone());
-    self.windows_api.set_window_placement_and_force_repaint(handle, placement);
-
-    // If margins are disabled, attempt a Desktop Window Manager-aware correction
-    if self.margin() == 0
-      && let Some(rect) = self
-        .windows_api
-        .get_extended_frame_bounds(handle)
-        .or_else(|| self.windows_api.get_window_rect(handle))
-      && let Some(compensating_rect) = calculate_compensating_rect_if_required(&rect, &sizing)
-    {
-      self.windows_api.set_window_position(handle, compensating_rect);
-    }
-  }
-
-  /// Returns the window under the cursor, if any. If there are multiple windows under the cursor, the foreground window
-  /// is returned if it's in the list. Otherwise, the window with the closest center point to the cursor is returned.
-  fn find_window_at_cursor<'a>(&self, point: &Point, windows: &'a [Window]) -> Option<&'a Window> {
-    let windows_under_cursor = windows
-      .iter()
-      .filter(|window_info| {
-        point.x() >= window_info.rect.left
-          && point.x() <= window_info.rect.right
-          && point.y() >= window_info.rect.top
-          && point.y() <= window_info.rect.bottom
-      })
-      .collect::<Vec<&Window>>();
-
-    if !windows_under_cursor.is_empty() {
-      if let Some(foreground_window) = self.windows_api.get_foreground_window()
-        && let Some(window_info) = windows_under_cursor
-          .iter()
-          .find(|window_info| window_info.handle == foreground_window)
-      {
-        debug!(
-          "Cursor is currently over foreground window {} \"{}\" at {point}",
-          window_info.handle,
-          window_info.title_trunc()
-        );
-        return Some(window_info);
-      }
-
-      let mut closest_window = None;
-      let mut min_distance = f64::MAX;
-      for window_info in windows_under_cursor {
-        let distance = ((window_info.center.x().pow(2) + window_info.center.y().pow(2)) as f64)
-          .sqrt()
-          .trunc();
-
-        if distance < min_distance {
-          min_distance = distance;
-          closest_window = Some(window_info);
-        }
-      }
-      let closest_window = closest_window.expect("Failed to get the closest window");
-      debug!(
-        "Cursor is currently over window {} \"{}\" at {point} with a distance of {}",
-        closest_window.handle,
-        closest_window.title_trunc(),
-        min_distance.trunc()
-      );
-      return Some(closest_window);
-    }
-
-    None
-  }
-
-  fn get_window_and_monitor_info(&self) -> Option<(WindowHandle, WindowPlacement, MonitorInfo)> {
-    let window = self.windows_api.get_foreground_window()?;
-    let placement = self.windows_api.get_window_placement(window)?;
-    let monitor_info = self.windows_api.get_monitor_info_for_window(window)?;
-    Some((window, placement, monitor_info))
-  }
-
-  fn move_focus_to_window(&self, direction: Direction, target_window: &Window, target_point: &Point) {
-    self.windows_api.set_cursor_position(target_point);
-    self.windows_api.set_foreground_window(WindowHandle::from(target_window));
-    info!(
-      "Moved cursor in direction [{:?}] to {} \"{}\" at {target_point}",
-      direction,
-      target_window.handle,
-      target_window.title_trunc()
-    );
-  }
-
-  fn move_focus_to_monitor(&self, direction: Direction, monitor: &Monitor) {
-    self.windows_api.set_cursor_position(&monitor.center);
-    info!(
-      "Moved cursor in direction [{:?}] to {} on [{}]",
-      direction, monitor.center, monitor
-    );
-  }
-
-  fn find_closest_window(&self, cursor_position: Point, ignored_window: Option<WindowHandle>) -> Option<WindowHandle> {
-    let mut closest_window = (vec![], f64::MAX);
-    self
-      .windows_api
-      .get_all_visible_windows()
-      .iter()
-      .filter(|w| ignored_window != Some(w.handle))
-      .for_each(|w| {
-        let distance = cursor_position.distance_to(&w.center);
-        trace!(
-          "Distance from cursor position {} to window {} \"{}\" is {}",
-          cursor_position,
-          w.handle,
-          w.title_trunc(),
-          distance
-        );
-        if distance == closest_window.1 {
-          closest_window.0.push(w.clone());
-        } else if distance < closest_window.1 {
-          closest_window.0.clear();
-          closest_window.0.push(w.clone());
-          closest_window.1 = distance;
-        }
-      });
-    match closest_window.0.len() {
-      0 => {
-        trace!("No windows found close to {}", cursor_position);
-
-        None
-      }
-      1 => Some(closest_window.0.first().map(|w| w.handle)?),
-      _ => {
-        let smallest_window = closest_window
-          .0
-          .iter()
-          .min_by_key(|window| {
-            let width = window.rect.right - window.rect.left;
-            let height = window.rect.bottom - window.rect.top;
-            width * height
-          })
-          .map(|window| window.handle)
-          .expect("Failed to get the smallest window");
-        trace!(
-          "Found multiple windows closest to cursor position {}, returning {} which is the smallest one",
-          cursor_position, smallest_window
-        );
-
-        Some(smallest_window)
-      }
-    }
-  }
-
-  fn find_and_select_closest_window(&mut self, window: WindowHandle) {
-    let cursor_position = self.windows_api.get_cursor_position();
-    if let Some(window) = self.find_closest_window(cursor_position, Some(window)) {
-      self.windows_api.set_foreground_window(window);
-      let window_info = self
-        .windows_api
-        .get_window_placement(window)
-        .expect("Failed to get window placement");
-      let target_point = Point::from_center_of_rect(&window_info.normal_position);
-      self.windows_api.set_cursor_position(&target_point);
-    } else {
-      info!("No window found to move focus to after closing the current window");
-    }
-  }
-
-  /// Determines whether the given window placement matches the expected sizing. If margins are disabled, allows a
-  /// small tolerance when comparing against the DWM extended frame bounds to account for shadows/rounded corners added
-  /// by the OS.
-  ///
-  /// Note: This extra check may be useful in all cases, but I don't understand the Windows API well enough yet, and
-  /// I've never had this problem before despite my heavy use of this application.
-  fn is_of_expected_size(&self, handle: WindowHandle, placement: &WindowPlacement, sizing: &Sizing) -> bool {
-    let rect = placement.normal_position;
-    let is_expected_size = rect.left == sizing.x
-      && rect.top == sizing.y
-      && rect.right - rect.left == sizing.width
-      && rect.bottom - rect.top == sizing.height;
-    if is_expected_size {
-      log_actual_vs_expected(&handle, sizing, rect);
-      debug!("{} is currently of expected size (exact placement match)", handle);
-      return true;
-    }
-
-    if self.margin() == 0
-      && let Some(compensating_rect) = self
-        .windows_api
-        .get_extended_frame_bounds(handle)
-        .or_else(|| self.windows_api.get_window_rect(handle))
-    {
-      let is_compensating_match = (compensating_rect.left - sizing.x).abs() <= DWM_TOLERANCE_IN_PX
-        && (compensating_rect.top - sizing.y).abs() <= DWM_TOLERANCE_IN_PX
-        && (compensating_rect.right - compensating_rect.left - sizing.width).abs() <= DWM_TOLERANCE_IN_PX
-        && (compensating_rect.bottom - compensating_rect.top - sizing.height).abs() <= DWM_TOLERANCE_IN_PX;
-      log_actual_vs_expected(&handle, sizing, compensating_rect);
-      debug!(
-        "{} {} of expected size (dwm_tolerance: {})",
-        handle,
-        if is_compensating_match {
-          "is currently"
-        } else {
-          "is currently NOT"
-        },
-        DWM_TOLERANCE_IN_PX
-      );
-
-      return is_compensating_match;
-    }
-
-    log_actual_vs_expected(&handle, sizing, rect);
-    debug!("{} is currently NOT of expected size (strict placement comparison)", handle,);
-    false
-  }
-}
-
-fn select_window_in_direction<'window>(
-  reference_point: &Point,
-  direction: Direction,
-  windows: &[&'window Window],
-  reference_window: Option<&Window>,
-  allow_selecting_same_center_windows: bool,
-) -> Option<&'window Window> {
-  // Cycle same-centre windows first
-  if allow_selecting_same_center_windows
-    && let Some(reference_window) = reference_window
-    && let Some(next_window) = find_next_same_center_window(reference_window, windows)
-  {
-    return Some(next_window);
-  }
-
-  let mut closest_window = None;
-  let mut closest_score = f64::MAX;
-
-  for &window in windows {
-    let target_center_x = window.rect.left + (window.rect.right - window.rect.left) / 2;
-    let target_center_y = window.rect.top + (window.rect.bottom - window.rect.top) / 2;
-    let dx = target_center_x as i64 - reference_point.x() as i64;
-    let dy = target_center_y as i64 - reference_point.y() as i64;
-    let should_filter_by_direction = !allow_selecting_same_center_windows
-      || reference_window.is_some_and(|reference_window| {
-        reference_window.center != window.center || reference_window.handle == window.handle
-      });
-
-    // Skip windows outside the requested direction
-    if should_filter_by_direction {
-      match direction {
-        Direction::Left if dx >= 0 => continue,
-        Direction::Right if dx <= 0 => continue,
-        Direction::Up if dy >= 0 => continue,
-        Direction::Down if dy <= 0 => continue,
-        _ => {}
-      }
-    }
-
-    // Score by distance and directional alignment
-    let distance = ((dx.pow(2) + dy.pow(2)) as f64).sqrt().trunc();
-    let angle = match direction {
-      Direction::Left => (dy as f64).atan2((-dx) as f64).abs(),
-      Direction::Right => (dy as f64).atan2(dx as f64).abs(),
-      Direction::Up => (dx as f64).atan2((-dy) as f64).abs(),
-      Direction::Down => (dx as f64).atan2(dy as f64).abs(),
-    };
-    let score = distance + angle;
-    trace!(
-      "Score for {} is [{}] (i.e. normalised_angle={}, distance={})",
-      window.handle,
-      score.trunc(),
-      angle,
-      distance,
-    );
-    if score < closest_score {
-      closest_score = score;
-      closest_window = Some(window);
-    }
-  }
-
-  closest_window
-}
-
-fn find_next_same_center_window<'window>(reference_window: &Window, windows: &[&'window Window]) -> Option<&'window Window> {
-  let mut same_center_windows = windows
-    .iter()
-    .copied()
-    .filter(|window| window.center == reference_window.center)
-    .collect::<Vec<_>>();
-  if same_center_windows.len() < 2 {
-    return None;
-  }
-
-  // Keep cycling independent of Z-order
-  same_center_windows.sort_unstable_by_key(|window| window.handle.hwnd);
-  let reference_index = same_center_windows
-    .iter()
-    .position(|window| window.handle == reference_window.handle)?;
-  same_center_windows
-    .get((reference_index + 1) % same_center_windows.len())
-    .copied()
-}
-
-fn interpolate_rect(start: Rect, target: Rect, progress: f64) -> Rect {
-  fn coordinate(start: i32, target: i32, progress: f64) -> i32 {
-    (f64::from(start) + f64::from(target - start) * progress.clamp(0.0, 1.0)).round() as i32
-  }
-
-  Rect::new(
-    coordinate(start.left, target.left, progress),
-    coordinate(start.top, target.top, progress),
-    coordinate(start.right, target.right, progress),
-    coordinate(start.bottom, target.bottom, progress),
-  )
-}
-
-fn add_or_update_previous_placement(
-  known_windows: &mut HashMap<String, WindowPlacement>,
-  handle: WindowHandle,
-  placement: WindowPlacement,
-) {
-  let window_id = format!("{:?}", handle.hwnd);
-  if known_windows.contains_key(&window_id) {
-    known_windows.remove(&window_id);
-    trace!(
-      "Removing previous placement for window {} so that a new value can be added",
-      handle
-    );
-  }
-
-  known_windows.insert(window_id.clone(), placement);
-  trace!("Adding/updating previous placement for window {}", handle);
-}
-
-fn log_actual_vs_expected(handle: &WindowHandle, sizing: &Sizing, rc: Rect) {
-  debug!(
-    "Expected size of {}: ({},{})x({},{})",
-    handle, sizing.x, sizing.y, sizing.width, sizing.height
-  );
-  debug!(
-    "Actual size of {}: ({},{})x({},{})",
-    handle,
-    rc.left,
-    rc.top,
-    rc.right - rc.left,
-    rc.bottom - rc.top
-  );
-}
-
-fn calculate_compensating_rect_if_required(rect: &Rect, sizing: &Sizing) -> Option<Rect> {
-  let requested_left = sizing.x;
-  let requested_right = sizing.x + sizing.width;
-  let left_inset = rect.left - requested_left;
-  let right_inset = requested_right - rect.right;
-
-  if left_inset > 0 || right_inset > 0 {
-    let compensating_left = requested_left - left_inset.max(0);
-    let compensating_right = requested_right + right_inset.max(0);
-    return Some(Rect::new(
-      compensating_left,
-      sizing.y,
-      compensating_right,
-      sizing.y + sizing.height,
-    ));
-  }
-  None
-}
-
-#[cfg(test)]
-mod tests {
-  use crate::api::{MockWindowsApi, WindowsApi};
-  use crate::common::{
-    Direction, MonitorHandle, PersistentWorkspaceId, Point, Rect, Sizing, Window, WindowHandle, WindowPlacement,
-  };
-  use crate::configuration_provider::{ConfigurationProvider, Layout};
-  use crate::horizontal_layout::HorizontalLayout;
-  use crate::utils::{MINIMUM_WINDOW_DIMENSION, MINIMUM_WINDOW_DIMENSION_DIVISOR, create_temp_directory};
-  use crate::window_manager::{DWM_TOLERANCE_IN_PX, WindowManager, select_window_in_direction};
-  use crate::workspace_manager::WorkspaceManager;
-  use std::collections::HashMap;
-  use std::path::PathBuf;
-  use std::sync::{Arc, Mutex};
-
-  impl WindowManager<MockWindowsApi> {
-    pub fn default(api: MockWindowsApi) -> Self {
-      WindowManager {
-        configuration_provider: Arc::new(Mutex::new(ConfigurationProvider::default())),
-        known_windows: Default::default(),
-        allow_moving_cursor_after_close_or_minimise: true,
-        horizontal_layout: HorizontalLayout::default(),
-        horizontal_layout_positions: HashMap::new(),
-        horizontal_layout_initialised: false,
-        previous_foreground_window: None,
-        workspace_manager: WorkspaceManager::default(),
-        virtual_desktop_manager: None,
-        windows_api: api,
-      }
-    }
-
-    pub fn new_test(api: MockWindowsApi, config_path: PathBuf) -> Self {
-      WindowManager {
-        configuration_provider: Arc::new(Mutex::new(ConfigurationProvider::new_test(config_path))),
-        known_windows: Default::default(),
-        allow_moving_cursor_after_close_or_minimise: true,
-        horizontal_layout: HorizontalLayout::default(),
-        horizontal_layout_positions: HashMap::new(),
-        horizontal_layout_initialised: false,
-        previous_foreground_window: None,
-        workspace_manager: WorkspaceManager::default(),
-        virtual_desktop_manager: None,
-        windows_api: api,
-      }
-    }
-  }
-
-  fn horizontal_manager() -> (WindowManager<MockWindowsApi>, tempfile::TempDir) {
-    MockWindowsApi::reset();
-    let directory = create_temp_directory();
-    let workspace_manager = WorkspaceManager::new_test(true, directory.path().join("workspaces.toml"));
-    let configuration_provider = Arc::new(Mutex::new(ConfigurationProvider::default()));
-    configuration_provider.lock().unwrap().set_default_layout(Layout::Scrolling);
-    let manager = WindowManager {
-      configuration_provider,
-      known_windows: Default::default(),
-      allow_moving_cursor_after_close_or_minimise: true,
-      horizontal_layout: HorizontalLayout::default(),
-      horizontal_layout_positions: HashMap::new(),
-      horizontal_layout_initialised: false,
-      previous_foreground_window: None,
-      workspace_manager,
-      virtual_desktop_manager: None,
-      windows_api: MockWindowsApi,
-    };
-    (manager, directory)
-  }
-
-  fn with_margin_set_to(margin: i32, manager: &mut WindowManager<MockWindowsApi>) {
-    manager
-      .configuration_provider
-      .lock()
-      .unwrap()
-      .set_i32(crate::configuration_provider::WINDOW_MARGIN, margin);
-  }
-
-  #[test]
-  fn reconciliation_only_manages_scrolling_monitors() {
-    MockWindowsApi::reset();
-    let directory = create_temp_directory();
-    let workspace_manager = WorkspaceManager::new_test(true, directory.path().join("workspaces.toml"));
-    let configuration_provider = Arc::new(Mutex::new(ConfigurationProvider::default()));
-    configuration_provider
-      .lock()
-      .unwrap()
-      .set_monitor_layout("DISPLAY1", Layout::Scrolling);
-    let secondary = WindowHandle::new(2);
-    let original = Sizing::new(-700, 50, 100, 100);
-    MockWindowsApi::add_or_update_window(secondary, "Secondary".to_string(), original.clone(), false, false, false);
-    MockWindowsApi::place_window(secondary, 2.into());
-    let mut manager = WindowManager {
-      configuration_provider,
-      known_windows: Default::default(),
-      allow_moving_cursor_after_close_or_minimise: true,
-      horizontal_layout: HorizontalLayout::default(),
-      horizontal_layout_positions: HashMap::new(),
-      horizontal_layout_initialised: false,
-      previous_foreground_window: None,
-      workspace_manager,
-      virtual_desktop_manager: None,
-      windows_api: MockWindowsApi,
-    };
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(
-      manager.windows_api.get_window_placement(1.into()).unwrap(),
-      WindowPlacement::new_from_sizing(Sizing::new(20, 20, 1880, 990))
-    );
-    assert_eq!(
-      manager.windows_api.get_window_placement(secondary).unwrap(),
-      WindowPlacement::new_from_sizing(original)
-    );
-    assert!(manager.horizontal_layout.workspace_containing(secondary).is_none());
-  }
-
-  #[test]
-  fn mixed_layout_routes_move_by_foreground_monitor() {
-    MockWindowsApi::reset();
-    let directory = create_temp_directory();
-    let workspace_manager = WorkspaceManager::new_test(true, directory.path().join("workspaces.toml"));
-    let configuration_provider = Arc::new(Mutex::new(ConfigurationProvider::default()));
-    configuration_provider
-      .lock()
-      .unwrap()
-      .set_monitor_layout("DISPLAY1", Layout::Scrolling);
-    let secondary = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      secondary,
-      "Secondary".to_string(),
-      Sizing::new(-700, 50, 400, 300),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::place_window(secondary, 2.into());
-    let mut manager = WindowManager {
-      configuration_provider,
-      known_windows: Default::default(),
-      allow_moving_cursor_after_close_or_minimise: true,
-      horizontal_layout: HorizontalLayout::default(),
-      horizontal_layout_positions: HashMap::new(),
-      horizontal_layout_initialised: false,
-      previous_foreground_window: None,
-      workspace_manager,
-      virtual_desktop_manager: None,
-      windows_api: MockWindowsApi,
-    };
-
-    manager.move_window(Direction::Up);
-    assert_eq!(
-      manager.windows_api.get_window_placement(secondary).unwrap(),
-      WindowPlacement::new_from_sizing(Sizing::top_half_of_screen(Rect::new(-800, 0, 0, 550), 20))
-    );
-    MockWindowsApi::set_foreground_window(1.into());
-    let primary_before = manager.windows_api.get_window_placement(1.into()).unwrap();
-    manager.move_window(Direction::Up);
-    assert_eq!(manager.windows_api.get_window_placement(1.into()).unwrap(), primary_before);
-  }
-
-  #[test]
-  fn moving_from_scrolling_to_spatial_removes_strip_membership() {
-    let (mut manager, _directory) = horizontal_manager();
-    manager
-      .configuration_provider
-      .lock()
-      .unwrap()
-      .set_monitor_layout("DISPLAY2", Layout::Spatial);
-    manager.reconcile_horizontal_layout();
-    let primary_workspace = PersistentWorkspaceId::from(*crate::workspace_manager::tests::primary_active_ws_id());
-    let secondary_workspace = manager
-      .workspace_manager
-      .active_workspace_ids()
-      .into_iter()
-      .find(|workspace| workspace.monitor_id != primary_workspace.monitor_id)
-      .unwrap();
-
-    manager.move_window_to_workspace(secondary_workspace);
-
-    assert!(manager.horizontal_layout.workspace_containing(1.into()).is_none());
-  }
-
-  #[test]
-  fn moving_from_spatial_to_scrolling_inserts_strip_membership() {
-    MockWindowsApi::reset();
-    let directory = create_temp_directory();
-    let workspace_manager = WorkspaceManager::new_test(false, directory.path().join("workspaces.toml"));
-    let configuration_provider = Arc::new(Mutex::new(ConfigurationProvider::default()));
-    configuration_provider
-      .lock()
-      .unwrap()
-      .set_monitor_layout("DISPLAY1", Layout::Scrolling);
-    let secondary = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      secondary,
-      "Secondary".to_string(),
-      Sizing::new(-700, 50, 400, 300),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::place_window(secondary, 2.into());
-    let primary_workspace = *crate::workspace_manager::tests::primary_active_ws_id();
-    let mut manager = WindowManager {
-      configuration_provider,
-      known_windows: Default::default(),
-      allow_moving_cursor_after_close_or_minimise: true,
-      horizontal_layout: HorizontalLayout::default(),
-      horizontal_layout_positions: HashMap::new(),
-      horizontal_layout_initialised: false,
-      previous_foreground_window: None,
-      workspace_manager,
-      virtual_desktop_manager: None,
-      windows_api: MockWindowsApi,
-    };
-    manager.reconcile_horizontal_layout();
-
-    manager.move_window_to_workspace(primary_workspace.into());
-
-    assert_eq!(
-      manager.horizontal_layout.workspace_containing(secondary),
-      Some(primary_workspace.into())
-    );
-  }
-
-  #[test]
-  fn disabled_horizontal_layout_does_not_reposition_windows() {
-    MockWindowsApi::reset();
-    let handle = WindowHandle::new(1);
-    let sizing = Sizing::new(5, 10, 100, 80);
-    MockWindowsApi::add_or_update_window(handle, "Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(1.into(), Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(handle, 1.into());
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(
-      manager.windows_api.get_window_placement(handle).unwrap(),
-      WindowPlacement::new_from_sizing(sizing)
-    );
-  }
-
-  #[test]
-  fn horizontal_reconciliation_ignores_non_manageable_windows() {
-    let (mut manager, _directory) = horizontal_manager();
-    let tray_menu = WindowHandle::new(2);
-    let original = Sizing::new(300, 300, 200, 100);
-    MockWindowsApi::add_or_update_window(tray_menu, "Tray menu".to_string(), original.clone(), false, false, false);
-    MockWindowsApi::place_window(tray_menu, 1.into());
-    MockWindowsApi::mark_window_unmanageable(tray_menu);
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(
-      manager.windows_api.get_window_placement(tray_menu).unwrap(),
-      WindowPlacement::new_from_sizing(original)
-    );
-  }
-
-  #[test]
-  fn horizontal_reconciliation_does_not_steal_focus_or_mouse_from_unmanaged_popup() {
-    let (mut manager, _directory) = horizontal_manager();
-    manager.reconcile_horizontal_layout();
-    let tray_popup = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      tray_popup,
-      "Tray popup".to_string(),
-      Sizing::new(1500, 900, 200, 100),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::place_window(tray_popup, 1.into());
-    MockWindowsApi::mark_window_unmanageable(tray_popup);
-    let click_position = Point::new(1800, 1000);
-    MockWindowsApi::set_cursor_position(click_position);
-    MockWindowsApi::clear_position_batches();
-
-    manager.reconcile_horizontal_layout();
-
-    assert!(MockWindowsApi::position_batches().is_empty());
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(tray_popup));
-    assert_eq!(manager.windows_api.get_cursor_position(), click_position);
-  }
-
-  #[test]
-  fn off_screen_member_keeps_its_stored_workspace_when_windows_reports_another_monitor() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-    MockWindowsApi::assign_window_to_monitor(second, 2.into());
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      1940
-    );
-  }
-
-  #[test]
-  fn unchanged_horizontal_reconciliation_does_not_reposition_windows() {
-    let (mut manager, _directory) = horizontal_manager();
-    manager.reconcile_horizontal_layout();
-    MockWindowsApi::clear_position_batches();
-
-    manager.reconcile_horizontal_layout();
-
-    assert!(MockWindowsApi::position_batches().is_empty());
-  }
-
-  #[test]
-  fn horizontal_reconciliation_does_not_reset_mouse_position() {
-    let (mut manager, _directory) = horizontal_manager();
-    manager.reconcile_horizontal_layout();
-    let mouse_position = Point::new(300, 400);
-    MockWindowsApi::set_cursor_position(mouse_position);
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(manager.windows_api.get_cursor_position(), mouse_position);
-  }
-
-  #[test]
-  fn horizontal_layout_adopts_and_places_existing_windows() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(
-      manager.windows_api.get_window_placement(1.into()).unwrap(),
-      WindowPlacement::new_from_sizing(Sizing::new(20, 20, 1880, 990))
-    );
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap(),
-      WindowPlacement::new_from_sizing(Sizing::new(1940, 20, 1880, 990))
-    );
-  }
-
-  #[test]
-  fn horizontal_layout_inserts_new_foreground_before_previous_foreground() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-    let new_window = WindowHandle::new(3);
-    MockWindowsApi::add_or_update_window(
-      new_window,
-      "New".to_string(),
-      Sizing::new(300, 50, 100, 100),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::place_window(new_window, 1.into());
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(new_window));
-    assert_eq!(
-      manager
-        .windows_api
-        .get_window_placement(new_window)
-        .unwrap()
-        .normal_position
-        .left,
-      20
-    );
-    assert_eq!(
-      manager
-        .windows_api
-        .get_window_placement(1.into())
-        .unwrap()
-        .normal_position
-        .left,
-      1940
-    );
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      3860
-    );
-  }
-
-  #[test]
-  fn horizontal_layout_focuses_new_window_even_before_it_becomes_foreground() {
-    let (mut manager, _directory) = horizontal_manager();
-    manager.reconcile_horizontal_layout();
-    let new_window = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      new_window,
-      "New".to_string(),
-      Sizing::new(300, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(new_window, 1.into());
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(new_window));
-    assert_eq!(
-      manager
-        .windows_api
-        .get_window_placement(new_window)
-        .unwrap()
-        .normal_position
-        .left,
-      20
-    );
-  }
-
-  #[test]
-  fn horizontal_focus_scrolls_strip_and_horizontal_move_reorders_it() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-
-    manager.move_cursor(Direction::Right);
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(second));
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      20
-    );
-
-    manager.move_window(Direction::Left);
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      20
-    );
-    assert_eq!(
-      manager
-        .windows_api
-        .get_window_placement(1.into())
-        .unwrap()
-        .normal_position
-        .left,
-      1940
-    );
-  }
-
-  #[test]
-  fn horizontal_focus_animates_batched_intermediate_positions() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-    MockWindowsApi::clear_position_batches();
-
-    manager.move_cursor(Direction::Right);
-
-    let batches = MockWindowsApi::position_batches();
-    assert!(batches.len() > 2);
-    let first_incoming = batches[0]
-      .iter()
-      .find(|(handle, _)| *handle == second)
-      .map(|(_, rect)| *rect)
-      .unwrap();
-    assert!(first_incoming.left < 1940 && first_incoming.left > 20);
-    assert_eq!(batches.last().unwrap()[0].0, second, "Focused window should be topmost");
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      20
-    );
-  }
-
-  #[test]
-  fn horizontal_layout_disables_vertical_move_and_resize() {
-    let (mut manager, _directory) = horizontal_manager();
-    manager.reconcile_horizontal_layout();
-    let initial = manager.windows_api.get_window_placement(1.into()).unwrap();
-
-    manager.move_window(Direction::Up);
-    manager.resize_window(Direction::Left);
-
-    assert_eq!(manager.windows_api.get_window_placement(1.into()).unwrap(), initial);
-  }
-
-  #[test]
-  fn horizontal_workspace_switch_preserves_off_screen_strip_members() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-    let first_workspace =
-      crate::common::PersistentWorkspaceId::from(*crate::workspace_manager::tests::primary_active_ws_id());
-    let second_workspace = manager
-      .workspace_manager
-      .workspaces
-      .keys()
-      .find(|id| id.monitor_id == first_workspace.monitor_id && id.workspace == 2)
-      .copied()
-      .unwrap();
-
-    manager.switch_workspace(second_workspace);
-    assert!(manager.windows_api.get_all_visible_windows().is_empty());
-    manager.switch_workspace(first_workspace);
-
-    assert_eq!(manager.windows_api.get_all_visible_windows().len(), 2);
-    assert_eq!(
-      manager
-        .windows_api
-        .get_window_placement(1.into())
-        .unwrap()
-        .normal_position
-        .left,
-      20
-    );
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      1940
-    );
-  }
-
-  #[test]
-  fn restoring_horizontal_layout_moves_off_screen_members_onto_their_monitor() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-    let first_workspace =
-      crate::common::PersistentWorkspaceId::from(*crate::workspace_manager::tests::primary_active_ws_id());
-    let empty_workspace = manager
-      .workspace_manager
-      .workspaces
-      .keys()
-      .find(|id| id.monitor_id == first_workspace.monitor_id && id.workspace == 2)
-      .copied()
-      .unwrap();
-    manager.switch_workspace(empty_workspace);
-
-    manager.restore_all_managed_windows();
-
-    let visible_windows = manager.windows_api.get_all_visible_windows();
-    assert_eq!(visible_windows.len(), 2);
-    assert!(visible_windows.iter().all(|window| {
-      window
-        .rect
-        .intersects(&crate::workspace_manager::tests::primary_monitor().monitor_area)
-    }));
-  }
-
-  #[test]
-  fn horizontal_move_to_workspace_updates_both_strips() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-    let first_workspace =
-      crate::common::PersistentWorkspaceId::from(*crate::workspace_manager::tests::primary_active_ws_id());
-    let second_workspace = manager
-      .workspace_manager
-      .workspaces
-      .keys()
-      .find(|id| id.monitor_id == first_workspace.monitor_id && id.workspace == 2)
-      .copied()
-      .unwrap();
-
-    manager.move_window_to_workspace(second_workspace);
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(second));
-    manager.switch_workspace(second_workspace);
-
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(1.into()));
-    assert_eq!(
-      manager
-        .windows_api
-        .get_window_placement(1.into())
-        .unwrap()
-        .normal_position
-        .left,
-      20
-    );
-  }
-
-  #[test]
-  fn horizontal_reconciliation_removes_externally_closed_member() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-    manager.windows_api.do_close_window(1.into());
-
-    manager.reconcile_horizontal_layout();
-
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(second));
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      20
-    );
-  }
-
-  #[test]
-  fn horizontal_layout_removes_closed_member_and_focuses_neighbour() {
-    let (mut manager, _directory) = horizontal_manager();
-    let second = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(
-      second,
-      "Second".to_string(),
-      Sizing::new(500, 50, 100, 100),
-      false,
-      false,
-      false,
-    );
-    MockWindowsApi::place_window(second, 1.into());
-    manager.reconcile_horizontal_layout();
-
-    manager.close_window();
-
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(second));
-    assert_eq!(
-      manager.windows_api.get_window_placement(second).unwrap().normal_position.left,
-      20
-    );
-  }
-
-  #[test]
-  fn is_of_expected_size_test() {
-    let handle = WindowHandle::new(1);
-    let placement = WindowPlacement::new_from_sizing(Sizing::new(0, 0, 100, 100));
-    let sizing = Sizing::new(0, 0, 100, 100);
-    let manager = WindowManager::default(MockWindowsApi);
-    assert!(manager.is_of_expected_size(handle, &placement, &sizing));
-
-    let placement = WindowPlacement::new_from_sizing(Sizing::new(1, 0, 101, 100));
-    assert!(!manager.is_of_expected_size(handle, &placement, &sizing));
-  }
-
-  #[test]
-  fn near_maximise_window_when_window_is_not_near_maximised() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::new(0, 0, 100, 100);
-    let initial_placement = WindowPlacement::new_from_sizing(sizing.clone());
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.near_maximise_or_restore();
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_placement = WindowPlacement::new_from_sizing(Sizing::new(20, 20, 160, 140));
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert!(manager.known_windows.contains_key(&format!("{:?}", window_handle.hwnd)));
-    assert_eq!(
-      *manager.known_windows.get(&format!("{:?}", window_handle.hwnd)).unwrap(),
-      initial_placement
-    );
-  }
-
-  #[test]
-  fn restore_window_when_window_is_near_maximised() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::new(20, 20, 160, 140);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-    let previous_placement = WindowPlacement::new_test();
-    manager
-      .known_windows
-      .insert(format!("{:?}", window_handle.hwnd), previous_placement.clone());
-
-    manager.near_maximise_or_restore();
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), previous_placement);
-  }
-
-  #[test]
-  fn move_window_on_the_same_monitor() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::new(20, 20, 160, 160);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mock_api = MockWindowsApi;
-    let mut manager = WindowManager::default(mock_api);
-
-    manager.move_window(Direction::Right);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let monitor_info = mock_api
-      .get_monitor_info_for_monitor(monitor_handle)
-      .expect("Failed to get monitor info");
-    let expected_sizing = Sizing::right_half_of_screen(monitor_info.work_area, 20);
-    let expected_cursor_position = Point::from_center_of_sizing(&expected_sizing);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing);
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(manager.windows_api.get_cursor_position(), expected_cursor_position)
-  }
-
-  #[test]
-  fn move_window_when_window_is_already_at_target_location() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::left_half_of_screen(Rect::new(0, 0, 200, 180), 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.move_window(Direction::Left);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_placement = WindowPlacement::new_from_sizing(sizing);
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(manager.windows_api.get_cursor_position(), Point::default())
-  }
-
-  #[test]
-  fn move_window_to_another_monitor() {
-    let monitor_handle_1 = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::right_half_of_screen(Rect::new(0, 0, 200, 180), 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle_1, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::add_monitor(2.into(), Rect::new(200, 0, 400, 200), false);
-    MockWindowsApi::place_window(window_handle, monitor_handle_1);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.move_window(Direction::Right);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_placement = WindowPlacement::new_from_sizing(Sizing::near_maximised(Rect::new(200, 0, 400, 180), 20));
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(manager.windows_api.get_cursor_position(), Point::new(300, 100))
-  }
-
-  #[test]
-  fn select_window_in_direction_cycles_through_all_windows() {
-    let rect = Rect::new(0, 0, 100, 100);
-    let first = Window::new_test(1, rect);
-    let second = Window::new_test(2, rect);
-    let third = Window::new_test(3, rect);
-    let windows = [&third, &first, &second];
-
-    assert_eq!(
-      select_window_in_direction(&first.center, Direction::Right, &windows, Some(&first), true).map(|window| window.handle),
-      Some(second.handle)
-    );
-    assert_eq!(
-      select_window_in_direction(&second.center, Direction::Right, &windows, Some(&second), true)
-        .map(|window| window.handle),
-      Some(third.handle)
-    );
-    assert_eq!(
-      select_window_in_direction(&third.center, Direction::Right, &windows, Some(&third), true).map(|window| window.handle),
-      Some(first.handle)
-    );
-  }
-
-  #[test]
-  fn select_window_in_direction_uses_direction_when_disabled() {
-    let reference = Window::new_test(1, Rect::new(0, 0, 100, 100));
-    let same_center = Window::new_test(2, Rect::new(0, 0, 100, 100));
-    let right = Window::new_test(3, Rect::new(100, 0, 200, 100));
-    let windows = [&reference, &same_center, &right];
-
-    let selected = select_window_in_direction(&reference.center, Direction::Right, &windows, Some(&reference), false);
-
-    assert_eq!(selected.map(|window| window.handle), Some(right.handle));
-  }
-
-  #[test]
-  fn select_window_in_direction_falls_back_to_closest_window_in_direction() {
-    let reference = Window::new_test(1, Rect::new(0, 0, 100, 100));
-    let closest_right = Window::new_test(2, Rect::new(100, 0, 200, 100));
-    let furthest_right = Window::new_test(3, Rect::new(200, 0, 300, 100));
-    let windows = [&reference, &furthest_right, &closest_right];
-
-    let selected = select_window_in_direction(&reference.center, Direction::Right, &windows, Some(&reference), true);
-
-    assert_eq!(selected.map(|window| window.handle), Some(closest_right.handle));
-  }
-
-  #[test]
-  fn move_cursor_moves_cursor_to_center_of_closest_window_on_other_monitor() {
-    let current_monitor_handle = MonitorHandle::from(1);
-    let target_monitor_handle = MonitorHandle::from(2);
-    let target_work_area = Rect::new(200, 0, 400, 200);
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::right_half_of_screen(target_work_area, 20);
-    MockWindowsApi::set_cursor_position(Point::new(0, 0));
-    MockWindowsApi::add_or_update_window(window_handle, "Test".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(current_monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::add_monitor(target_monitor_handle, target_work_area, true);
-    MockWindowsApi::place_window(window_handle, target_monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.move_cursor(Direction::Right);
-
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(window_handle));
-    assert_eq!(manager.windows_api.get_cursor_position(), target_work_area.center());
-  }
-
-  #[test]
-  fn move_cursor_does_nothing_when_there_is_no_window_or_monitor_to_move_to() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle_1 = WindowHandle::new(1);
-    let left_sizing = Sizing::left_half_of_screen(Rect::new(0, 0, 200, 180), 20);
-    let initial_cursor_position = Point::new(0, 0);
-    MockWindowsApi::set_cursor_position(initial_cursor_position);
-    MockWindowsApi::add_or_update_window(window_handle_1, "Test".to_string(), left_sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle_1, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.move_cursor(Direction::Right);
-
-    assert_eq!(manager.windows_api.get_cursor_position(), initial_cursor_position);
-  }
-
-  #[test]
-  fn close_window_does_close_window() {
-    let window_handle = WindowHandle::new(1);
-    let monitor_handle = MonitorHandle::from(1);
-    MockWindowsApi::add_or_update_window(window_handle, "Test".to_string(), Sizing::default(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.close_window();
-
-    assert!(
-      !manager
-        .windows_api
-        .get_all_visible_windows()
-        .iter()
-        .any(|w| w.handle == window_handle)
-    );
-  }
-
-  #[test]
-  fn close_window_moves_cursor_to_closest_window_when_enabled() {
-    let window_handle_1 = WindowHandle::new(1);
-    let window_handle_2 = WindowHandle::new(2);
-    let monitor_handle = MonitorHandle::from(1);
-    MockWindowsApi::add_or_update_window(window_handle_1, "Test 1".to_string(), Sizing::default(), false, false, true);
-    let sizing = Sizing::new(0, 0, 100, 100);
-    MockWindowsApi::add_or_update_window(window_handle_2, "Test 2".to_string(), sizing, false, false, false);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle_1, monitor_handle);
-    MockWindowsApi::place_window(window_handle_2, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.close_window();
-
-    assert!(
-      !manager
-        .windows_api
-        .get_all_visible_windows()
-        .iter()
-        .any(|w| w.handle == window_handle_1)
-    );
-    assert_eq!(manager.windows_api.get_cursor_position(), Point::new(50, 50));
-  }
-
-  #[test]
-  fn close_window_does_not_move_cursor_to_closest_window_when_disabled() {
-    let window_handle_1 = WindowHandle::new(1);
-    let window_handle_2 = WindowHandle::new(2);
-    let monitor_handle = MonitorHandle::from(1);
-    MockWindowsApi::add_or_update_window(window_handle_1, "Test 1".to_string(), Sizing::default(), false, false, true);
-    let sizing = Sizing::new(0, 0, 100, 100);
-    MockWindowsApi::add_or_update_window(window_handle_2, "Test 2".to_string(), sizing, false, false, false);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle_1, monitor_handle);
-    MockWindowsApi::place_window(window_handle_2, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-    manager.allow_moving_cursor_after_close_or_minimise = false;
-
-    manager.close_window();
-
-    assert!(
-      !manager
-        .windows_api
-        .get_all_visible_windows()
-        .iter()
-        .any(|w| w.handle == window_handle_1)
-    );
-    assert_eq!(manager.windows_api.get_cursor_position(), Point::default());
-  }
-
-  #[test]
-  fn close_window_fails_silently() {
-    let mut manager = WindowManager::default(MockWindowsApi);
-    assert!(manager.windows_api.get_all_visible_windows().is_empty());
-
-    manager.close_window();
-
-    assert!(manager.windows_api.get_all_visible_windows().is_empty());
-  }
-
-  #[test]
-  fn find_closest_window_returns_none_when_no_windows_are_visible() {
-    let cursor_position = Point::new(100, 100);
-    let manager = WindowManager::default(MockWindowsApi);
-
-    let result = manager.find_closest_window(cursor_position, None);
-
-    assert!(result.is_none());
-  }
-
-  #[test]
-  fn find_closest_window_returns_window_under_cursor() {
-    let cursor_position = Point::new(50, 50);
-    let window_handle = WindowHandle::new(1);
-    MockWindowsApi::set_cursor_position(cursor_position);
-    MockWindowsApi::add_or_update_window(
-      window_handle,
-      "Test Window".to_string(),
-      Rect::new(0, 0, 100, 100).into(),
-      false,
-      false,
-      true,
-    );
-    let manager = WindowManager::default(MockWindowsApi);
-
-    let result = manager.find_closest_window(cursor_position, None);
-
-    assert_eq!(result, Some(window_handle));
-  }
-
-  #[test]
-  fn find_closest_window_returns_smallest_window_when_multiple_windows_overlap() {
-    let cursor_position = Point::new(50, 50);
-    let expected_window_handle = WindowHandle::new(2);
-    MockWindowsApi::set_cursor_position(cursor_position);
-    MockWindowsApi::add_or_update_window(
-      WindowHandle::new(1),
-      "Window 1".to_string(),
-      Rect::new(40, 40, 60, 60).into(),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::add_or_update_window(
-      expected_window_handle,
-      "Window 2".to_string(),
-      Rect::new(45, 45, 55, 55).into(),
-      false,
-      false,
-      false,
-    );
-    let manager = WindowManager::default(MockWindowsApi);
-
-    let result = manager.find_closest_window(cursor_position, None);
-
-    assert_eq!(result, Some(expected_window_handle));
-  }
-
-  #[test]
-  fn find_closest_window_returns_closest_window_when_cursor_is_outside_all_windows() {
-    let cursor_position = Point::new(1000, 1000);
-    let window_handle = WindowHandle::new(1);
-    MockWindowsApi::set_cursor_position(cursor_position);
-    MockWindowsApi::add_or_update_window(
-      window_handle,
-      "Test Window".to_string(),
-      Rect::new(0, 0, 100, 100).into(),
-      false,
-      false,
-      false,
-    );
-    let manager = WindowManager::default(MockWindowsApi);
-
-    let result = manager.find_closest_window(cursor_position, None);
-
-    assert_eq!(result, Some(window_handle));
-  }
-
-  #[test]
-  fn find_closest_window_ignores_minimised_or_hidden_windows() {
-    let cursor_position = Point::new(50, 50);
-    let expected_window_handle = WindowHandle::new(2);
-    MockWindowsApi::set_cursor_position(cursor_position);
-    MockWindowsApi::add_or_update_window(
-      WindowHandle::new(1),
-      "Window 1".to_string(),
-      Rect::new(40, 40, 60, 60).into(),
-      true,
-      false,
-      true,
-    );
-    MockWindowsApi::add_or_update_window(
-      expected_window_handle,
-      "Window 2".to_string(),
-      Rect::new(45, 45, 55, 55).into(),
-      false,
-      true,
-      false,
-    );
-    let manager = WindowManager::default(MockWindowsApi);
-
-    let result = manager.find_closest_window(cursor_position, None);
-
-    assert!(result.is_none());
-  }
-
-  #[test]
-  fn find_closest_window_ignores_provided_window() {
-    let cursor_position = Point::new(50, 50);
-    let expected_window_handle = WindowHandle::new(2);
-    MockWindowsApi::set_cursor_position(cursor_position);
-    MockWindowsApi::add_or_update_window(
-      WindowHandle::new(1),
-      "Window 1".to_string(),
-      Rect::new(0, 0, 60, 60).into(),
-      true,
-      false,
-      true,
-    );
-    let manager = WindowManager::default(MockWindowsApi);
-
-    let result = manager.find_closest_window(cursor_position, Some(expected_window_handle));
-
-    assert!(result.is_none());
-  }
-
-  #[test]
-  fn minimise_window_when_no_other_window_present() {
-    let window_handle = WindowHandle::new(1);
-    MockWindowsApi::add_or_update_window(window_handle, "Test".to_string(), Sizing::default(), false, false, true);
-    let mut manager = WindowManager::default(MockWindowsApi);
-    assert!(!manager.windows_api.is_window_minimised(window_handle));
-
-    manager.minimise_window();
-
-    assert!(manager.windows_api.is_window_minimised(window_handle));
-    assert_eq!(manager.windows_api.get_cursor_position(), Point::default());
-    assert_eq!(manager.windows_api.get_foreground_window(), None);
-  }
-
-  #[test]
-  fn minimise_window_when_another_window_is_present() {
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::new(50, 50, 100, 100);
-    let other_window_handle = WindowHandle::new(2);
-    MockWindowsApi::add_or_update_window(window_handle, "Test".to_string(), Sizing::default(), false, false, true);
-    MockWindowsApi::add_or_update_window(other_window_handle, "Other".to_string(), sizing.clone(), false, false, false);
-    let mut manager = WindowManager::default(MockWindowsApi);
-    assert!(!manager.windows_api.is_window_minimised(window_handle));
-
-    manager.minimise_window();
-
-    assert!(manager.windows_api.is_window_minimised(window_handle));
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&sizing),
-      "Cursor position should be set to the center of the closest, non-minimised window"
-    );
-    assert_eq!(manager.windows_api.get_foreground_window(), Some(other_window_handle));
-  }
-
-  #[test]
-  fn near_maximise_window_with_margin_below_threshold_does_not_resize() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let initial_sizing = Sizing::new(50, 50, 100, 100);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), initial_sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let manager = WindowManager::default(MockWindowsApi);
-    let monitor_info = manager
-      .windows_api
-      .get_monitor_info_for_monitor(monitor_handle)
-      .expect("Failed to get monitor info");
-
-    manager.near_maximise_window(window_handle, monitor_info, 3);
-
-    let expected_sizing = Sizing::near_maximised(monitor_info.work_area, 0);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-  }
-
-  #[test]
-  fn near_maximise_window_with_margin_above_threshold_resizes() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let initial_sizing = Sizing::new(50, 50, 100, 100);
-    let initial_placement = WindowPlacement::new_from_sizing(initial_sizing.clone());
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), initial_sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let manager = WindowManager::default(MockWindowsApi);
-    let monitor_info = manager
-      .windows_api
-      .get_monitor_info_for_monitor(monitor_handle)
-      .expect("Failed to get monitor info");
-    let margin = 10;
-
-    manager.near_maximise_window(window_handle, monitor_info, margin);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = Sizing::near_maximised(monitor_info.work_area, margin);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing);
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.clone().unwrap(), expected_placement);
-    assert_ne!(actual_placement.unwrap(), initial_placement);
-  }
-
-  #[test]
-  fn near_maximise_or_restore_with_zero_margin_can_restore_initial_position() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let initial_sizing = Sizing::new(50, 50, 100, 100);
-    let initial_placement = WindowPlacement::new_from_sizing(initial_sizing.clone());
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), initial_sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 200, 200), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-    let monitor_info = manager
-      .windows_api
-      .get_monitor_info_for_monitor(monitor_handle)
-      .expect("Failed to get monitor info");
-    with_margin_set_to(0, &mut manager);
-
-    manager.near_maximise_or_restore();
-
-    let expected_sizing = Sizing::near_maximised(monitor_info.work_area, 0);
-    let maximised_placement = WindowPlacement::new_from_sizing(expected_sizing);
-    let current_placement = manager
-      .windows_api
-      .get_window_placement(window_handle)
-      .expect("Failed to get placement after maximise");
-    assert_eq!(current_placement, maximised_placement, "Window should be maximised");
-
-    manager.near_maximise_or_restore();
-
-    let current_placement = manager
-      .windows_api
-      .get_window_placement(window_handle)
-      .expect("Failed to get placement after restore");
-    assert_eq!(
-      current_placement, initial_placement,
-      "Window should be restored to its initial placement"
-    );
-  }
-
-  #[test]
-  fn is_of_expected_size_returns_true_when_same_size() {
-    let directory = create_temp_directory();
-    let path = directory.path().join("file.toml");
-    let mut manager = WindowManager::new_test(MockWindowsApi, path);
-    let window_handle = WindowHandle::new(1);
-    let rect = Rect::new(0, 0, 1280, 720);
-    let placement = WindowPlacement::new_from_rect(rect);
-    let sizing = rect.into();
-    with_margin_set_to(10, &mut manager);
-
-    assert!(manager.is_of_expected_size(window_handle, &placement, &sizing));
-  }
-
-  #[test]
-  fn is_of_expected_size_returns_false_when_different_size() {
-    let directory = create_temp_directory();
-    let path = directory.path().join("file.toml");
-    let mut manager = WindowManager::new_test(MockWindowsApi, path);
-    let window_handle = WindowHandle::new(1);
-    let rect = Rect::new(0, 0, 1280, 720);
-    let placement = WindowPlacement::new_from_rect(rect);
-    let mut sizing: Sizing = rect.into();
-    sizing.x += 5; // Introducing a discrepancy here
-    with_margin_set_to(10, &mut manager);
-
-    assert!(!manager.is_of_expected_size(window_handle, &placement, &sizing));
-  }
-
-  #[test]
-  fn is_of_expected_size_returns_true_when_difference_is_within_dwm_tolerance() {
-    let directory = create_temp_directory();
-    let path = directory.path().join("file.toml");
-    let window_handle = WindowHandle::new(1);
-    let rect = Rect::new(0, 0, 1280, 720);
-    let placement = WindowPlacement::new_from_rect(rect);
-    let expected_sizing = Sizing::new(0, 0, 640, 720);
-    // Windows API returns a different sizing that is within the tolerance for a no-margin configuration
-    let api_sizing = Sizing::new(DWM_TOLERANCE_IN_PX - 1, DWM_TOLERANCE_IN_PX - 1, 640, 720);
-    MockWindowsApi::add_or_update_window(
-      window_handle,
-      "Test Window".to_string(),
-      api_sizing.clone(),
-      false,
-      false,
-      true,
-    );
-    let mut manager = WindowManager::new_test(MockWindowsApi, path);
-    with_margin_set_to(0, &mut manager);
-
-    assert!(manager.is_of_expected_size(window_handle, &placement, &expected_sizing));
-  }
-
-  #[test]
-  fn is_of_expected_size_returns_false_when_difference_is_outside_dwm_tolerance() {
-    let directory = create_temp_directory();
-    let path = directory.path().join("file.toml");
-    let window_handle = WindowHandle::new(1);
-    let rect = Rect::new(0, 0, 1280, 720);
-    let placement = WindowPlacement::new_from_rect(rect);
-    let expected_sizing = Sizing::new(0, 0, 640, 720);
-    // Windows API returns a different sizing that is outside the tolerance for a no-margin configuration
-    let api_sizing = Sizing::new(15, 0, 640, 720);
-    MockWindowsApi::add_or_update_window(
-      window_handle,
-      "Test Window".to_string(),
-      api_sizing.clone(),
-      false,
-      false,
-      true,
-    );
-    let mut manager = WindowManager::new_test(MockWindowsApi, path);
-    with_margin_set_to(0, &mut manager);
-
-    assert!(!manager.is_of_expected_size(window_handle, &placement, &expected_sizing));
-  }
-
-  #[test]
-  fn resize_window_steps_near_maximised_down_to_three_quarter_in_direction() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    // Monitor area bottom is 20px more than work_area bottom (mock subtracts 20 for taskbar)
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::near_maximised(work_area, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Left, 20);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing.clone());
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_halves_non_near_maximised_window_in_direction() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    // Width must be > 2 * dynamic_min (2 * W/4 = 1000) so that halved width (590) > dynamic_min (500)
-    let sizing = Sizing::new(20, 20, 1200, 960);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = sizing.halved(Direction::Left, 20);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing.clone());
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_steps_near_maximised_down_to_three_quarter_down() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::near_maximised(work_area, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Down);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Down, 20);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing.clone());
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_steps_three_quarter_left_down_to_left_half_of_screen() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Left, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = Sizing::left_half_of_screen(work_area, 20);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing.clone());
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_steps_three_quarter_down_down_to_bottom_half_of_screen() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Down, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Down);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = Sizing::bottom_half_of_screen(work_area, 20);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing.clone());
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_three_quarter_left_halves_normally_in_down_direction() {
-    // A 75%-wide window (from Left resize) should NOT trigger the 3/4 rule when pressing Down
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Left, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Down);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = sizing.halved(Direction::Down, 20);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing.clone());
-    assert!(actual_placement.is_some());
-    assert_eq!(actual_placement.unwrap(), expected_placement);
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_does_nothing_when_dynamic_minimum_exceeds_constant_and_result_is_below_it() {
-    // Use a screen wide enough that W/DIVISOR > MINIMUM_WINDOW_DIMENSION, so the dynamic minimum
-    // takes over. Work area width = MINIMUM * DIVISOR * 2, giving dynamic_min = MINIMUM * 2.
-    // The window is sized so its halved width falls between the constant and the dynamic minimum.
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area_width = MINIMUM_WINDOW_DIMENSION * MINIMUM_WINDOW_DIMENSION_DIVISOR * 2;
-    // dynamic_min = work_area_width / DIVISOR = MINIMUM_WINDOW_DIMENSION * 2
-    let sizing = Sizing::new(20, 20, MINIMUM_WINDOW_DIMENSION * 3, 600);
-    // halved width = (MINIMUM * 3) / 2 - half_margin (with default margin 20)
-    // = 375 - 10 = 365, which is > MINIMUM (250) but < dynamic_min (500) → blocked
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, work_area_width, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let initial_cursor = Point::default();
-    MockWindowsApi::set_cursor_position(initial_cursor);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(sizing),
-      "Window should not have been resized"
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      initial_cursor,
-      "Cursor should not have moved"
-    );
-  }
-
-  #[test]
-  fn resize_window_allows_resize_when_constant_is_larger_than_quarter_screen() {
-    // Dynamic min = max(MINIMUM_WINDOW_DIMENSION, work_area/DIVISOR). On a 1000px-wide screen,
-    // work_area/DIVISOR = 980/8 = 122 < MINIMUM_WINDOW_DIMENSION (250), so the constant wins.
-    // A window whose halved width (365px) exceeds the constant should be allowed to resize.
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    // Monitor 1000px wide -> work_area 1000px wide -> quarter = 250 < MINIMUM_WINDOW_DIMENSION (350)
-    let sizing = Sizing::new(20, 20, 750, 560);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 1000, 620), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_sizing = sizing.halved(Direction::Left, 20);
-    let expected_placement = WindowPlacement::new_from_sizing(expected_sizing.clone());
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      expected_placement,
-      "Window should have been resized"
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_halves_left_half_of_screen_in_left_direction() {
-    // With DIVISOR=8, dynamic_min = max(MINIMUM, W/8). On a 2000px screen, dynamic_min = max(250, 250) = 250.
-    // Halving left_half produces ~475px > 250 -> resize succeeds.
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, MINIMUM_WINDOW_DIMENSION * MINIMUM_WINDOW_DIMENSION_DIVISOR, 1000);
-    let left_half = Sizing::left_half_of_screen(work_area, 20);
-    MockWindowsApi::add_or_update_window(
-      window_handle,
-      "Test Window".to_string(),
-      left_half.clone(),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::add_monitor(
-      monitor_handle,
-      Rect::new(0, 0, MINIMUM_WINDOW_DIMENSION * MINIMUM_WINDOW_DIMENSION_DIVISOR, 1020),
-      true,
-    );
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let initial_cursor = Point::default();
-    MockWindowsApi::set_cursor_position(initial_cursor);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let expected = left_half.halved(Direction::Left, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected),
-      "Window should have been halved"
-    );
-    assert_ne!(
-      manager.windows_api.get_cursor_position(),
-      initial_cursor,
-      "Cursor should have moved"
-    );
-  }
-
-  #[test]
-  fn resize_window_halves_right_half_of_screen_in_right_direction() {
-    // Mirror of the Left case: with DIVISOR=8, dynamic_min = max(250, 250) = 250.
-    // Halving right_half produces ~475px > 250 -> resize succeeds.
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, MINIMUM_WINDOW_DIMENSION * MINIMUM_WINDOW_DIMENSION_DIVISOR, 1000);
-    let right_half = Sizing::right_half_of_screen(work_area, 20);
-    MockWindowsApi::add_or_update_window(
-      window_handle,
-      "Test Window".to_string(),
-      right_half.clone(),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::add_monitor(
-      monitor_handle,
-      Rect::new(0, 0, MINIMUM_WINDOW_DIMENSION * MINIMUM_WINDOW_DIMENSION_DIVISOR, 1020),
-      true,
-    );
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let initial_cursor = Point::default();
-    MockWindowsApi::set_cursor_position(initial_cursor);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Right);
-
-    let expected = right_half.halved(Direction::Right, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected),
-      "Window should have been halved"
-    );
-    assert_ne!(
-      manager.windows_api.get_cursor_position(),
-      initial_cursor,
-      "Cursor should have moved"
-    );
-  }
-
-  #[test]
-  fn resize_window_does_nothing_when_below_minimum_dimension() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let small_sizing = Sizing::new(20, 20, 200, 200);
-    MockWindowsApi::add_or_update_window(
-      window_handle,
-      "Test Window".to_string(),
-      small_sizing.clone(),
-      false,
-      false,
-      true,
-    );
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1000), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let initial_cursor = Point::default();
-    MockWindowsApi::set_cursor_position(initial_cursor);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    let expected_placement = WindowPlacement::new_from_sizing(small_sizing);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      expected_placement,
-      "Window should not have been resized"
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      initial_cursor,
-      "Cursor should not have moved"
-    );
-  }
-
-  #[test]
-  fn resize_window_does_nothing_when_no_foreground_window() {
-    // No window is added or focused -> get_window_and_monitor_info returns None -> early return.
-    let initial_cursor = Point::default();
-    MockWindowsApi::set_cursor_position(initial_cursor);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      initial_cursor,
-      "Cursor should not have moved"
-    );
-  }
-
-  #[test]
-  fn resize_window_does_nothing_when_result_height_falls_below_constant_minimum() {
-    // Down halved height = 300/2 - half_margin = 140 < MINIMUM_WINDOW_DIMENSION (250) -> blocked.
-    // On this screen, dynamic_min_height = max(250, 980/8) = 250, so the constant is the threshold.
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let sizing = Sizing::new(20, 20, 500, 300);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1000), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let initial_cursor = Point::default();
-    MockWindowsApi::set_cursor_position(initial_cursor);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Down);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(sizing),
-      "Window should not have been resized"
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      initial_cursor,
-      "Cursor should not have moved"
-    );
-  }
-
-  #[test]
-  fn resize_window_does_nothing_when_dynamic_minimum_height_blocks_resize() {
-    // Work area height = MINIMUM * DIVISOR * 2 -> dynamic_min_height = MINIMUM * 2.
-    // Window height = MINIMUM * 3; halved = MINIMUM * 3 / 2 - half_margin, which is
-    // > MINIMUM (constant) but < dynamic_min (MINIMUM * 2) -> blocked.
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area_height = MINIMUM_WINDOW_DIMENSION * MINIMUM_WINDOW_DIMENSION_DIVISOR * 2;
-    let sizing = Sizing::new(20, 20, 500, MINIMUM_WINDOW_DIMENSION * 3);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing.clone(), false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, work_area_height + 20), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let initial_cursor = Point::default();
-    MockWindowsApi::set_cursor_position(initial_cursor);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Down);
-
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(sizing),
-      "Window should not have been resized"
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      initial_cursor,
-      "Cursor should not have moved"
-    );
-  }
-
-  #[test]
-  fn resize_window_steps_near_maximised_down_to_three_quarter_right() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::near_maximised(work_area, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Right);
-
-    let expected_sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Right, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_steps_near_maximised_down_to_three_quarter_up() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::near_maximised(work_area, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Up);
-
-    let expected_sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Up, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_steps_three_quarter_right_down_to_right_half_of_screen() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Right, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Right);
-
-    let expected_sizing = Sizing::right_half_of_screen(work_area, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_steps_three_quarter_up_down_to_top_half_of_screen() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Up, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Up);
-
-    let expected_sizing = Sizing::top_half_of_screen(work_area, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_three_quarter_left_produces_centre_when_pressing_right() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Left, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Right);
-
-    let expected_sizing = Sizing::centre_near_maximised(work_area, Direction::Right, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_three_quarter_right_produces_centre_when_pressing_left() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Right, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Left);
-
-    let expected_sizing = Sizing::centre_near_maximised(work_area, Direction::Left, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_three_quarter_up_produces_centre_when_pressing_down() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Up, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Down);
-
-    let expected_sizing = Sizing::centre_near_maximised(work_area, Direction::Down, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-
-  #[test]
-  fn resize_window_three_quarter_down_produces_centre_when_pressing_up() {
-    let monitor_handle = MonitorHandle::from(1);
-    let window_handle = WindowHandle::new(1);
-    let work_area = Rect::new(0, 0, 2000, 1000);
-    let sizing = Sizing::three_quarter_near_maximised(work_area, Direction::Down, 20);
-    MockWindowsApi::add_or_update_window(window_handle, "Test Window".to_string(), sizing, false, false, true);
-    MockWindowsApi::add_monitor(monitor_handle, Rect::new(0, 0, 2000, 1020), true);
-    MockWindowsApi::place_window(window_handle, monitor_handle);
-    let mut manager = WindowManager::default(MockWindowsApi);
-
-    manager.resize_window(Direction::Up);
-
-    let expected_sizing = Sizing::centre_near_maximised(work_area, Direction::Up, 20);
-    let actual_placement = manager.windows_api.get_window_placement(window_handle);
-    assert!(actual_placement.is_some());
-    assert_eq!(
-      actual_placement.unwrap(),
-      WindowPlacement::new_from_sizing(expected_sizing.clone())
-    );
-    assert_eq!(
-      manager.windows_api.get_cursor_position(),
-      Point::from_center_of_sizing(&expected_sizing)
-    );
-  }
-}
