+++ src/window_manager/tests/mod.rs
@@ -1,0 +1,8 @@
+use crate::window_manager::window_manager;
+
+mod navigation_tests;
+mod placement_tests;
+mod scrolling_tests;
+mod spatial_tests;
+mod test_support;
+mod window_manager_tests;
